+++ app/src/main/java/edu/javeriana/fixup/ui/features/publication_detail/PublicationDetailViewModel.kt
@@ -71,7 +71,7 @@ class PublicationDetailViewModel @Inject constructor(
             val request = ReviewRequestDto(
                 userId = 1, 
                 serviceId = serviceId,
-                rating = rating.toDouble(),
+                rating = rating,
                 comment = comment
             )
